✨ feat: 모집글 상세 조회 기능에 조회수 기능 추가

+++ src/main/java/efub/SweetMeback/domain/post/dto/PostResponseDto.java
@@ -22,7 +22,7 @@ public class PostResponseDto {
     private LocalDateTime startDate;
     private LocalDateTime endDate;
     private Integer people;
-    private Long view;
+    private Integer view;
     private boolean recruitment;
     private Category category;
     private Meeting meeting;
